move management services from shared to manage folder
