+++ src/main/webapp/app/tutorialgroup/manage/tutorial-groups-checklist/tutorial-groups-checklist.component.spec.ts
@@ -11,7 +11,7 @@ import { AlertService } from 'app/shared/service/alert.service';
 import { mockedActivatedRoute } from 'test/helpers/mocks/activated-route/mock-activated-route-query-param-map';
 import { HttpResponse } from '@angular/common/http';
 import { of } from 'rxjs';
-import { TutorialGroupsConfigurationService } from 'app/tutorialgroup/shared/service/tutorial-groups-configuration.service';
+import { TutorialGroupsConfigurationService } from 'app/tutorialgroup/manage/service/tutorial-groups-configuration.service';
 import { generateExampleTutorialGroupsConfigurationDTO } from 'test/helpers/sample/tutorialgroup/tutorialGroupsConfigurationExampleModels';
 import { MockTranslateService } from 'test/helpers/mocks/service/mock-translate.service';
 import { TranslateService } from '@ngx-translate/core';
