+++ src/main/webapp/app/tutorialgroup/manage/tutorial-groups/tutorial-groups-management/tutorial-groups-management.component.spec.ts
@@ -16,7 +16,7 @@ import { mockedActivatedRoute } from 'test/helpers/mocks/activated-route/mock-ac
 import { By } from '@angular/platform-browser';
 import { generateExampleTutorialGroupsConfiguration } from 'test/helpers/sample/tutorialgroup/tutorialGroupsConfigurationExampleModels';
 import { Course } from 'app/core/course/shared/entities/course.model';
-import { TutorialGroupsConfigurationService } from 'app/tutorialgroup/shared/service/tutorial-groups-configuration.service';
+import { TutorialGroupsConfigurationService } from 'app/tutorialgroup/manage/service/tutorial-groups-configuration.service';
 import { TutorialGroupsImportButtonComponent } from './tutorial-groups-import-button/tutorial-groups-import-button.component';
 import { TutorialGroupsExportButtonComponent } from './tutorial-groups-export-button.component/tutorial-groups-export-button.component';
 import { MockTranslateService } from 'test/helpers/mocks/service/mock-translate.service';
