+++ src/main/webapp/app/tutorialgroup/shared/tutorial-group-sessions-table/tutorial-group-session-row/tutorial-group-session-row.component.ts
@@ -27,7 +27,7 @@ import { TranslateDirective } from 'app/shared/language/translate.directive';
 import { NgTemplateOutlet } from '@angular/common';
 import { ArtemisDatePipe } from 'app/shared/pipes/artemis-date.pipe';
 import { ArtemisTranslatePipe } from 'app/shared/pipes/artemis-translate.pipe';
-import { TutorialGroupSessionService } from 'app/tutorialgroup/shared/service/tutorial-group-session.service';
+import { TutorialGroupSessionService } from 'app/tutorialgroup/manage/service/tutorial-group-session.service';
 
 @Component({
     // this is intended and an attribute selector because otherwise the rendered table breaks
